% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{as.MS}
\alias{as.MS}
\title{Coerce to MS.}
\usage{
as.MS(x)
}
\arguments{
\item{x}{a MassSpectrum Java-Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Functions to check if an object is a MSM, or coerce it if possible.
}
