% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoperold.R
\name{as.MSM}
\alias{as.MSM}
\title{Coerce to MSM.}
\usage{
as.MSM(x)
}
\arguments{
\item{x}{a MassSpectrometryMeasurement Java-Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Functions to check if an object is a MSM, or coerce it if possible.
}
