% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{CreateProperties}
\alias{CreateProperties}
\title{Creates a properties file as config.}
\usage{
CreateProperties(filename)
}
\arguments{
\item{filename}{}
}
\description{
Creates a properties file as config.
}
\author{
Lucas Schmidt
}
