% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{deisotoper}
\alias{deisotoper}
\alias{getConfig}
\title{Creates a deisotoper.}
\usage{
deisotoper(AA_MASS = c(71.03711, 156.10111, 114.04293, 115.02694, 103.00919,
  129.04259, 128.05858, 57.02146, 137.05891, 113.08406, 113.08406, 128.09496,
  131.04049, 147.06841, 97.05276, 87.03203, 101.04768, 186.07931, 163.06333,
  99.06841), F1 = 0.8, F2 = 0.5, F3 = 0.1, F4 = 0.1, F5 = 0.1,
  DELTA = 0.003, ERRORTOLERANCE = 0.3, DISTANCE = 1.003, NOISE = 0,
  DECHARGE = FALSE)
}
\arguments{
\item{AA_MASS}{Vector of amino acid masses used for scoring}

\item{F1}{F1 multiplier used for scoring}

\item{F2}{F2 multiplier used for scoring}

\item{F3}{F3 multiplier used for scoring}

\item{F4}{F4 multiplier used for scoring}

\item{F5}{F5 multiplier used for scoring}

\item{DELTA}{Delta value used for clustering}

\item{ERRORTOLERANCE}{Errortolerance used for scoring}

\item{DISTANCE}{Distance between two peaks used by clustering}

\item{NOISE}{Noise value for noise filtering (in percent)}

\item{DECHARGE}{De- and activates decharging}
}
\value{
deisotoper
}
\description{
Creates a deisotoper.
}
\examples{
# standart configurated deisotoper
dtoper <- deisotoper()

config <- getConfig(dtoper)

# standart configurated deisotoper with changed delta and decharging
dtoper2 <- deisotoper(DELTA = 0.005, DECHARGE = TRUE)

config2 <- getConfig(dtoper2)
}
\author{
Lucas Schmidt
}
