% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoperold.R
\name{is.MS}
\alias{is.MS}
\title{Coerce to MS.}
\usage{
is.MS(x)
}
\arguments{
\item{x}{a MassSpectrum Java-Object}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Functions to check if an object is a MS, or coerce it if possible.
}
