% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateConfiguration}
\alias{jCreateConfiguration}
\title{Creates a Configuration (Java Object).}
\usage{
jCreateConfiguration(aa = c(), F1 = 0.8, F2 = 0.5, F3 = 0.1, F4 = 0.1,
  F5 = 0.1, DELTA = 0.003, ERRORTOLERANCE = 0.3, DISTANCE = 1.003,
  NOISE = 0, DECHARGE = FALSE)
}
\arguments{
\item{aa}{Amino Acid Masses Vector}

\item{F1}{Multiplier F1}

\item{F2}{Multiplier F2}

\item{F3}{Multiplier F3}

\item{F4}{Multiplier F4}

\item{F5}{Multiplier F5}

\item{DELTA}{Delta for clustering}

\item{ERRORTOLERANCE}{Errortolerance for scoring}

\item{DISTANCE}{Distance between two peaks}

\item{NOISE}{Noise filtering}

\item{DECHARGE}{Logical parameter, if decharge or not}
}
\value{
Configuration
}
\description{
Creates a Configuration (Java Object).
}
\author{
Lucas Schmidt
}
