% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateIMS}
\alias{jCreateIMS}
\title{Creates a isotopicmassspectrum from a massspectrum}
\usage{
jCreateIMS(massspectrum, configfile = "nofile")
}
\arguments{
\item{massspectrum}{a MassSpectrum Java-Object}

\item{configfile}{properties file}
}
\value{
ims
}
\description{
Creates a isotopicmassspectrum from a massspectrum
}
\author{
Lucas Schmidt
}
