% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateMS}
\alias{jCreateMS}
\title{Input loaded Obj with [i] (example: obj[1]).}
\usage{
jCreateMS(obj)
}
\arguments{
\item{obj}{}
}
\value{
ms
}
\description{
Input loaded Obj with [i] (example: obj[1]).
}
\author{
Lucas Schmidt
}
