% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeconvoluteMSMmspy}
\alias{jDeconvoluteMSMmspy}
\title{Deconvolutes a MSM.}
\usage{
jDeconvoluteMSMmspy(jobj)
}
\description{
Deconvolutes a MSM.
}
\examples{

load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

joMSMdeisotoped <- jDeisotopeMSMmspy(joMSM)

joMSMsummary <- jSummaryMSM(joMSM)
joMSMdeisotopedsummary <- jSummaryMSM(joMSMdeisotoped)

joMSMdeconvoluted <- jDeconvoluteMSMmspy(joMSMdeisotoped)
joMSMdeconvolutedsummary <- jSummaryMSM(joMSMdeconvoluted)
}
