% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeisotopeMS}
\alias{jDeisotopeMS}
\title{Deisotopes a MS.}
\usage{
jDeisotopeMS(ms, modus = "first", configfile = "nofile")
}
\description{
Deisotopes a MS.
}
\examples{

load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

mslist<-joMSM$getMSlist()
ms <- mslist$get(as.integer(0)) #index 0 of joMSM

msdeisotoped <- jDeisotopeMS(ms)

}
