% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeisotopeMSfdbm}
\alias{jDeisotopeMSfdbm}
\title{Deisotopes a MS.}
\usage{
jDeisotopeMSfdbm(ms, save = FALSE, modus = "first", aamassfile = "nofile",
  percent = 0, errortolerance = 0.3, delta = 0.01)
}
\description{
Deisotopes a MS.
}
\examples{

load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

mslist<-joMSM$getMSlist()
ms <- mslist$get(as.integer(0)) #index 0 of joMSM

msdeisotoped <- jDeisotopeMSfdbm(ms)

}
