% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeisotopeMSM}
\alias{jDeisotopeMSM}
\title{Deisotopes a MSM. Save and first parameters are for method fdbm.}
\usage{
jDeisotopeMSM(jobj, method = "fdbm", save = FALSE, modus = "first",
  aamassfile = "nofile", percent = 0, errortolerance = 0.3,
  delta = 0.01)
}
\description{
Deisotopes a MSM. Save and first parameters are for method fdbm.
}
\examples{
load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

joMSMdeisotoped <- jDeisotopeMSM(joMSM)

joMSMsummary <- jSummaryMSM(joMSM)
joMSMdeisotopedsummary <- jSummaryMSM(joMSMdeisotoped)

}
