% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeisotopeMSMfdbm}
\alias{jDeisotopeMSMfdbm}
\title{Deisotopes a MSM.}
\usage{
jDeisotopeMSMfdbm(jobj, save = FALSE, modus = "first",
  aamassfile = "nofile", percent = 0, errortolerance = 0.3,
  delta = 0.01)
}
\description{
Deisotopes a MSM.
}
\examples{

load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

joMSMdeisotoped <- jDeisotopeMSMfdbm(joMSM)

joMSMsummary <- jSummaryMSM(joMSM)
joMSMdeisotopedsummary <- jSummaryMSM(joMSMdeisotoped)

}
