% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jGetDot}
\alias{jGetDot}
\title{Creates the dot graph from the given massspectrum and the given isotopicset.}
\usage{
jGetDot(massspectrum, isotopicmassspectrum, index, configuration)
}
\arguments{
\item{massspectrum}{a IsotopicMassSpectrum Java-Object}

\item{isotopicmassspectrum}{a MassSpectrum Java-Object}

\item{index}{index of IsotopicSet}

\item{configfile}{properties file}
}
\value{
dot graph
}
\description{
Creates the dot graph from the given massspectrum and the given isotopicset.
}
\author{
Lucas Schmidt
}
