% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jGetDotGraphFromIS}
\alias{jGetDotGraphFromIS}
\title{Wrapper function to get a dot graph from a single isotopicset.}
\usage{
jGetDotGraphFromIS(massspectrum, isotopicset, configfile = "nofile")
}
\arguments{
\item{massspectrum}{}

\item{isotopicset}{}

\item{configfile}{}
}
\value{
dot graph
}
\description{
Wrapper function to get a dot graph from a single isotopicset.
}
\author{
Lucas Schmidt
}
