% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jGetIS}
\alias{jGetIS}
\title{Returns a single isotopicset.}
\usage{
jGetIS(isotopicmassspectrum, index)
}
\arguments{
\item{isotopicmassspectrum}{}

\item{index}{}
}
\value{
is
}
\description{
Returns a single isotopicset.
}
\author{
Lucas Schmidt
}
