% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoperold.R
\name{jGetMS}
\alias{jGetMS}
\title{Gets the MS from Java to R.}
\usage{
jGetMS(massspectrum)
}
\arguments{
\item{massspectrum}{a MassSpectrum Java-Object}
}
\description{
Gets the MS from Java to R.
}
\author{
Lucas Schmidt
}
