% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jGetMSM}
\alias{jGetMSM}
\title{Gets the MSM from Java to R.}
\usage{
jGetMSM(jobj)
}
\arguments{
\item{jobj}{a MassSpectrometryMeasurement Java-Object}
}
\description{
Gets the MSM from Java to R.
}
\examples{
load(system.file("extdata",
name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)
}
\author{
Lucas Schmidt
}
