% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoperold.R
\name{jReadJSON2MSM}
\alias{jReadJSON2MSM}
\title{Deserializes a json into a MSM.}
\usage{
jReadJSON2MSM(filename = "test.json")
}
\arguments{
\item{filename}{from where it should deserialized}
}
\description{
Deserializes a json into a MSM.
}
\author{
Lucas Schmidt
}
