% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoperold.R
\name{jReadMGF}
\alias{jReadMGF}
\title{Reads a MGF and creates a MSM object from the input.}
\usage{
jReadMGF(filename)
}
\arguments{
\item{filename}{mgf}
}
\value{
msm
}
\description{
Reads a MGF and creates a MSM object from the input.
}
\author{
Lucas Schmidt
}
