% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jScoreConfigAsCSV}
\alias{jScoreConfigAsCSV}
\title{Returns the ScoreConfig which is in the properties file as a CSV.}
\usage{
jScoreConfigAsCSV(file)
}
\arguments{
\item{file}{properties file}
}
\value{
csv
}
\description{
Returns the ScoreConfig which is in the properties file as a CSV.
}
\author{
Lucas Schmidt
}
