% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jSummaryMS}
\alias{jSummaryMS}
\title{Makes summary from one massspectrum}
\usage{
jSummaryMS(jobj)
}
\arguments{
\item{jobj}{a MassSpectrum Java-Object}
}
\value{
csv
}
\description{
Makes summary from one massspectrum
}
\author{
Lucas Schmidt
}
