% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jSummaryMSM}
\alias{jSummaryMSM}
\title{Summary of a MSM object}
\usage{
jSummaryMSM(jobj)
}
\arguments{
\item{jobj}{a Java object reference generated by \code{\link{jCreateMSM}}.}
}
\value{
a \code{\link{data.frame}}
}
\description{
Summary of a MSM object
}
\examples{

 mzXMLfilename <- system.file("extdata",
      name='20161010_04_TP_HeLa_200ng.mzXML', package = "deisotoper")
      
 jo <- jCreateMSM(as.psmSet.mzXML(openMSfile(mzXMLfilename)))
 X <- jSummaryMSM(jo)
 
library(lattice)

xyplot(Value ~ SpectrumID | Attribute,
 data=X, 
 pch='.')
}
\author{
Lucas Schmidt, Witold E. Wolski, Christian Panse
}
