% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jWriteMGF}
\alias{jWriteMGF}
\title{Writes a MGF from a given MSM object.}
\usage{
jWriteMGF(filename, msm)
}
\arguments{
\item{filename}{}

\item{msm}{}
}
\value{
status
}
\description{
Writes a MGF from a given MSM object.
}
\author{
Lucas Schmidt
}
