% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jWriteMSM2JSON}
\alias{jWriteMSM2JSON}
\title{Serializes a MSM into a json.}
\usage{
jWriteMSM2JSON(jobj, filename = "test.json")
}
\arguments{
\item{filename}{}
}
\description{
Serializes a MSM into a json.
}
\examples{

}
