% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{mgf}
\alias{mgf}
\title{write Mascot Generic File (MGF)}
\usage{
mgf(obj, filename = paste(tempfile(), "mgf", sep = "."))
}
\arguments{
\item{obj}{}

\item{filename}{}
}
\value{
\code{file.info(filename)}
}
\description{
write Mascot Generic File (MGF)
}
