% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{AddToProperties}
\alias{AddToProperties}
\title{Key can be: Any Amino Acid (Short form, 1 or 3 letter-code), F1, F2, F3, F4, F5, DISTANCE, DELTA, ERRORTOLERANCE, NOISE}
\usage{
AddToProperties(filename, key, value)
}
\arguments{
\item{filename}{}

\item{key}{}

\item{value}{}
}
\description{
Key can be: Any Amino Acid (Short form, 1 or 3 letter-code), F1, F2, F3, F4, F5, DISTANCE, DELTA, ERRORTOLERANCE, NOISE
}
