% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzXML.R
\name{as.psmSet.mzXML}
\alias{as.psmSet.mzXML}
\title{convert mzXML into psmSet object}
\usage{
\method{as.psmSet}{mzXML}(obj)
}
\arguments{
\item{x}{a mzXML object}
}
\value{
a psmSet object
}
\description{
convert mzXML into psmSet object
}
\examples{

\dontrun{
 library(msdata); library(mzR)
  mzXML <- openMSfile("~/data/20161010_04_TP_HeLa_200ng.mzXML")
}
}
