% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzXML.R
\name{as.psmSet.mzXML}
\alias{as.psmSet.mzXML}
\title{convert mzXML into psmSet object}
\usage{
as.psmSet.mzXML(obj)
}
\arguments{
\item{obj}{a mzXML object}
}
\value{
a psmSet object
}
\description{
convert mzXML into psmSet object
}
\examples{
\dontrun{
 mzXMLfilename <- system.file("extdata",
      name='20161010_04_TP_HeLa_200ng.mzXML', package = "deisotoper")
      
 library(mzR)
 jo <- jCreateMSM(as.psmSet.mzXML(openMSfile(mzXMLfilename)))
 
 library(lattice)
  xyplot(Value ~ SpectrumID | Attribute, 
     data = jSummaryMSM(jo), 
     scales = list(relation = "sliced"), 
     pch=22)
} 
}
