% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{deisotope}
\alias{deisotope}
\title{Deisotope a Mass Spectrum}
\usage{
deisotope(deisotoper, massspectrum, method = "features-based")
}
\arguments{
\item{deisotoper}{a \code{\link{deisotoper}} object.}

\item{massspectrum}{a list of numeric vectors \code{mZ} and \code{intensity}.}

\item{method}{the supported deisotope algorithm, default is \code{method="features-based"}.}
}
\description{
Deisotope a Mass Spectrum
}
\references{
Features-Based Deisotoping Method for Tandem Mass Spectra \url{http://dx.doi.org/10.1155/2011/210805}.
}
\seealso{
\code{\link{deisotoper}}
}
\author{
Lucas Schmidt
}
