% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jBenchmark}
\alias{jBenchmark}
\title{Used for benchmarking. The input is a MGF and the output is also a MGF.}
\usage{
jBenchmark(input, output, method = "fdbm", save = FALSE, modus = "first",
  aamassfile = "nofile", percent = 0, errortolerance = 0.3,
  delta = 0.01)
}
\description{
Used for benchmarking. The input is a MGF and the output is also a MGF.
}
