% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jBenchmark}
\alias{jBenchmark}
\title{Used for benchmarking. The input is a RData and the output is also a MGF.}
\usage{
jBenchmark(input, output, modus = "first",
  configuration = jCreateConfiguration())
}
\arguments{
\item{input}{RData}

\item{output}{mgf}

\item{modus}{modus of aggregation}

\item{configfile}{properties file}
}
\value{
mgf
}
\description{
Used for benchmarking. The input is a RData and the output is also a MGF.
}
\author{
Lucas Schmidt
}
