% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateMSM}
\alias{jCreateMSM}
\title{Constructor for a Java object of type \code{MassSpectrometryMeasurement}}
\usage{
jCreateMSM(obj)
}
\arguments{
\item{obj}{a \code{\link{psmSet}} object}
}
\value{
returns a jcall object reference for a MSM
}
\description{
Constructor for a Java object of type \code{MassSpectrometryMeasurement}
}
\examples{
\dontrun{
 urlrdata <- "http://fgcz-ms.uzh.ch/~cpanse/data/386248.RData"
 con <- url(urlrdata)
 load(con)
 # library(protViz)
 S <- as.psmSet(F225712)
 jo <- jCreateMSM(S)
}
}
\author{
Lucas Schmidt
}
