% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateMSM}
\alias{jCreateMSM}
\title{creates Mass Spectrometry Measurement (MSM) Java object}
\usage{
jCreateMSM(obj)
}
\arguments{
\item{obj}{\code{protViz""psmSet} object}
}
\value{
returns a jcall object reference
}
\description{
creates Mass Spectrometry Measurement (MSM) Java object
}
\examples{
\dontrun{
 urlrdata = "http://fgcz-ms.uzh.ch/~cpanse/data/386248.RData"
 con <- url(urlrdata)
 load(con)
 S <- as.psmSet(F225712)
 jo <- jCreateMSM(S)
}
}
\author{
Lucas Schmidt, Witold E. Wolski, Christian Panse
}
