% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateMSM}
\alias{jCreateMSM}
\title{Serialize a set of data}
\usage{
jCreateMSM(obj)
}
\arguments{
\item{obj}{\code{\link{psmSet}} object}
}
\value{
returns a jcall object reference
}
\description{
Serialize a set of data
}
\examples{
\dontrun{
 urlrdata = "http://fgcz-ms.uzh.ch/~cpanse/data/386248.RData"
 con <- url(urlrdata)
 load(con)
 S <- as.psmSet(F225712)
 jo <- jCreateMSM(S)
}
}
