% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jCreateMSM}
\alias{jCreateMSM}
\title{creates Mass Spectrometry Measurement (MSM) Java object .}
\usage{
jCreateMSM(obj)
}
\arguments{
\item{obj}{\code{protViz""psmSet} object}
}
\value{
returns a jcall object reference
}
\description{
creates Mass Spectrometry Measurement (MSM) Java object .
}
\examples{
load(system.file("extdata",
   name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)


\dontrun{
 mzXMLfilename <- system.file("extdata",
      name='20161010_04_TP_HeLa_200ng.mzXML', package = "deisotoper")
      
 jo <- jCreateMSM(as.psmSet.mzXML(openMSfile(mzXMLfilename)))
}
}
\seealso{
\code{\link{jSummaryMSM}}
}
\author{
Lucas Schmidt, Witold E. Wolski, Christian Panse
}
