% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeisotopeMSM}
\alias{jDeisotopeMSM}
\title{Deisotopes a MSM.}
\usage{
jDeisotopeMSM(jobj, modus = "first", configuration = jCreateConfiguration())
}
\arguments{
\item{jobj}{a MassSpectrometryMeasurement Java-Object}

\item{modus}{modus of aggregation}

\item{configfile}{properties file}
}
\value{
deisotoped MSM
}
\description{
Deisotopes a MSM.
}
\examples{
load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

joMSMdeisotoped <- jDeisotopeMSM(joMSM)

joMSMsummary <- jSummaryMSM(joMSM)
joMSMdeisotopedsummary <- jSummaryMSM(joMSMdeisotoped)
}
\author{
Lucas Schmidt
}
