% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jDeisotopeMSMfdbm}
\alias{jDeisotopeMSMfdbm}
\title{Deisotopes a MSM.}
\usage{
jDeisotopeMSMfdbm(jobj, save = FALSE, modus = "first",
  aamassfile = "AminoAcidMasses.ini")
}
\description{
The amino acid masses file ("AminoAcidMasses.ini") must be in the home directory, otherwise it will initialize the standart amino acid masses.
}
\examples{

load(system.file("extdata", name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)

joMSMdeisotoped <- jDeisotopeMSMfdbm(joMSM)

joMSMsummary <- jSummaryMSM(joMSM)
joMSMdeisotopedsummary <- jSummaryMSM(joMSMdeisotoped)

}
