% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoperold.R
\name{jGetDotGraphFromIS}
\alias{jGetDotGraphFromIS}
\title{Wrapper function to get a dot graph from a single isotopicset.}
\usage{
jGetDotGraphFromIS(massspectrum, isotopicset,
  configuration = jCreateConfiguration())
}
\arguments{
\item{massspectrum}{a MassSpectrum Java-Object}

\item{isotopicset}{a IsotopicSet Java-Object}

\item{configfile}{properties file}
}
\value{
dot graph
}
\description{
Wrapper function to get a dot graph from a single isotopicset.
}
\author{
Lucas Schmidt
}
