% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jGetMSM}
\alias{jGetMSM}
\title{Title}
\usage{
jGetMSM(jobj)
}
\arguments{
\item{jobj}{}
}
\description{
Title
}
\examples{
load(system.file("extdata",
name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)
}
