% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jGetStatisticMSM}
\alias{jGetStatisticMSM}
\title{Returns a statistic as CSV}
\usage{
jGetStatisticMSM(msm, configfile = "nofile")
}
\arguments{
\item{msm}{a MassSpectrometryMeasurement Java-Object}

\item{configfile}{properties file}
}
\value{
csv
}
\description{
Returns a statistic as CSV
}
\author{
Lucas Schmidt
}
