% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jReadJSON2MSM}
\alias{jReadJSON2MSM}
\title{Deserializes a json into a MSM.}
\usage{
jReadJSON2MSM(filename = "test.json")
}
\arguments{
\item{filename}{}
}
\description{
Deserializes a json into a MSM.
}
\examples{

}
