% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jSummaryMSM}
\alias{jSummaryMSM}
\title{Summary of a MSM object}
\usage{
jSummaryMSM(jobj)
}
\arguments{
\item{jobj}{}
}
\value{
a \code{\link{data.frame}}
}
\description{
Summary of a MSM object
}
\examples{

\dontrun{
urlrdata = "http://fgcz-ms.uzh.ch/~cpanse/data/386248.RData"
con <- url(urlrdata)
load(con)
S <- as.psmSet(F225712)
jo <- jCreateMSM(S)
X <- jSummaryMSM(jo)

library(lattice)

xyplot(Value ~ SpectrumID | Attribute,
 data=X, 
 scales = list(y=list(log=TRUE)),
 pch='.')
}
}
\author{
Lucas Schmidt, Witold E. Wolski, Christian Panse
}
