% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jSummaryMSM}
\alias{jSummaryMSM}
\title{Summary of a MSM object}
\usage{
jSummaryMSM(jobj)
}
\arguments{
\item{jobj}{a Java object reference generated by \code{\link{jCreateMSM}}.}
}
\value{
a \code{\link{data.frame}}
}
\description{
Summary of a MSM object
}
\examples{
load(system.file("extdata",
   name='TP_HeLa_200ng_filtered_pd21.RData', package = "deisotoper"))
 joMSM <- jCreateMSM(TP_HeLa_200ng_filtered_pd21)
 X <- jSummaryMSM(joMSM)
 
par(mfrow=c(3,4)); 
rv <- lapply(unique(X$Attribute), 
 function(a){
   hist(X[X$Attribute == a, 3],xlab=a, main='Histgram')
   })

if(require(lattice)){

xyplot(Value ~ SpectrumID | Attribute,
 data=X, 
 pch='.')

 xyplot(Value ~ SpectrumID | Attribute, data=X, scales = list(y=list(log=TRUE)))
   
   
 histogram(~Value | Attribute, data=X)
 histogram(~Value|Attribute, data=X, scales = list(x=list(log=TRUE)))
 histogram(~Value|Attribute, data=X, scales = list(x=list(log=TRUE)), type='count')

 bwplot(~Value | Attribute, data=X, scales = list(x = list(log = TRUE)))
 }
}
\seealso{
\code{\link{jCreateMSM}}
}
\author{
Lucas Schmidt, Witold E. Wolski, Christian Panse
}
