% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deisotoper.R
\name{jWriteMSM2JSON}
\alias{jWriteMSM2JSON}
\title{writes a MSM java object into a json file}
\usage{
jWriteMSM2JSON(jobj, filename = "test.json")
}
\arguments{
\item{jobj}{a \code{\link{jCreateMSM}} generated object reference.}

\item{filename}{of the json file}
}
\value{
NULL jWriteMSM2JSON
}
\description{
writes a MSM java object into a json file
}
\examples{
\dontrun{
 urlrdata <- "http://fgcz-ms.uzh.ch/~cpanse/data/386248.RData"
 con <- url(urlrdata)
 load(con)
 # library(protViz)
 S <- as.psmSet(F225712)
 jo <- jCreateMSM(S)
 jWriteMSM2JSON(jo, filename)
}
}
\author{
Lucas Schmidt
}
